+++ Utils.cpp
@@ -1,4 +1,11 @@
+#include <string>
+#include <iostream>
+#include <clocale>
+#include <locale>
+#include <vector>
+
 #include "Utils.h"
+
 using namespace irr::core;
 using namespace irr::scene;
 using namespace irr::video;
@@ -57,6 +64,91 @@ std::wstring Utility::parentOfPath(const wstring &path)
         }
     }
     return ret;
+}
+
+wstring Utility::basename(const wstring &path)
+{
+    std::wstring ret = path;
+    std::wstring::size_type lastSlashPos = path.find_last_of(L"/");
+    if (lastSlashPos == std::wstring::npos) {
+        std::wstring::size_type lastSlashPos = path.find_last_of(L"\\");
+    }
+    if (lastSlashPos != std::wstring::npos) {
+        ret = path.substr(lastSlashPos+1);
+    }
+    return ret;
+}
+
+wstring Utility::withoutExtension(const wstring &path)
+{
+    std::wstring ret = path;
+    std::wstring::size_type lastDotPos = path.find_last_of(L".");
+    if (lastDotPos != std::wstring::npos) {
+        std::wstring::size_type lastSlashPos = path.find_last_of(L"/");
+        if (lastSlashPos == std::wstring::npos) {
+            std::wstring::size_type lastSlashPos = path.find_last_of(L"\\");
+        }
+        if (lastSlashPos != std::wstring::npos) {
+            if (lastDotPos > lastSlashPos) ret = path.substr(0, lastDotPos);
+        }
+        else ret = path.substr(0, lastDotPos);
+    }
+    return ret;
+}
+
+wstring Utility::delimiter(const wstring &path)
+{
+    std::wstring ret = L"/";
+    std::wstring::size_type lastSlashPos = path.find_last_of(L"/");
+    if (lastSlashPos == std::wstring::npos) {
+        // ret = "/";
+    }
+    else {
+        std::wstring::size_type lastSlashPos = path.find_last_of(L"\\");
+        if (lastSlashPos != std::wstring::npos) {
+            ret = L"\\";
+        }
+    }
+    return ret;
+}
+
+bool Utility::exists(const std::string& name) {
+    if (FILE *file = fopen(name.c_str(), "r")) {
+        fclose(file);
+        return true;
+    } else {
+        return false;
+    }
+}
+
+std::string Utility::toString(const std::wstring& ws) {
+    std::string ret;
+    // convert to w_string using locale: see Phillipp on <https://stackoverflow.com/questions/4804298/how-to-convert-wstring-into-string>
+    std::setlocale(LC_ALL, "");
+    const std::locale locale("");
+    typedef std::codecvt<wchar_t, char, std::mbstate_t> converter_type;
+    const converter_type& converter = std::use_facet<converter_type>(locale);
+    std::vector<char> to(ws.length() * converter.max_length());
+    std::mbstate_t state;
+    const wchar_t* from_next;
+    char* to_next;
+    const converter_type::result result = converter.out(state, ws.data(), ws.data() + ws.length(), from_next, &to[0], &to[0] + to.size(), to_next);
+    if (result == converter_type::ok or result == converter_type::noconv) {
+        const std::string s(&to[0], to_next);
+        std::cout <<"std::string =     "<<s<<std::endl;
+        ret += s;
+    }
+    return ret;
+}
+
+bool Utility::exists(const std::wstring& name) {
+    std::string name_s = toString(name);
+    if (FILE *file = fopen(name_s.c_str(), "r")) {
+        fclose(file);
+        return true;
+    } else {
+        return false;
+    }
 }