+++ Utils.h
@@ -11,6 +11,12 @@ public:
     static void dumpVectorToConsole( const irr::core::vector3df &vector );
     static void dumpMeshInfoToConsole( irr::scene::IAnimatedMeshSceneNode *node );
     static std::wstring parentOfPath(const std::wstring &path);
+    static std::wstring basename(const std::wstring &path);
+    static std::wstring withoutExtension(const std::wstring &path);
+    static std::wstring delimiter(const std::wstring &path);
+    static bool exists(const std::string &name);
+    static bool exists(const std::wstring &name);
+    static std::string toString(const std::wstring &name);
 };
 
 #endif // UTILS_H